==> Notebooks/Rmd/P2-Análisis-Explanatorio.Rmd <==
---
title: "Analisis Explanatorio"
output:
  html_document:
    df_print: paged
---
```{r}

# Load the data
tfg <- read.csv("datos_tfg.csv")
#datatypes
for (col in names(tfg)) {
  if (is.character(tfg[[col]])) {
    tfg[[col]] <- as.factor(tfg[[col]])
  }
}
tfg$SQL <- as.factor(tfg$SQL)
tfg$python <- as.factor(tfg$python)
tfg$java <- as.factor(tfg$java)
tfg$c_plus_plus <- as.factor(tfg$c_plus_plus)
tfg$cloud <- as.factor(tfg$cloud)
tfg$Country <- NULL
tfg$Employment <- as.factor(tfg$Employment)
tfg$Employed <- as.factor(tfg$Employed)
tfg$HaveWorkedWith <- as.character(tfg$HaveWorkedWith)
str(tfg)
```

```{r}
summary(tfg)
```

```{r}
# tablas de distribución
tabla <- list()

# Loop sobre cada columna del dataframe
for (col in names(tfg)) {
  # Verificar si la columna es un factor
  if (is.factor(tfg[[col]])) {
    # Calcular la distribución de frecuencia en porcentaje
    porcentaje <- prop.table(table(tfg[[col]])) * 100
    
    # Agregar la tabla de distribución a la lista
    tabla[[col]] <- as.data.frame(porcentaje)
  }
}

# Visualizar las tablas de distribución
print(tabla)
```

```{r}
#piechart para todas las variables 
# Obtener nombres de las variables que son factores
factor_vars <- names(tfg)[sapply(tfg, is.factor)]

# Crear una lista para almacenar los gráficos de pastel
pie_charts <- list()

# Iterar sobre cada variable que es factor
for (var in factor_vars) {
  # Contar las frecuencias de las categorías
  var_counts <- table(tfg[[var]])
  
  # Crear un dataframe con los datos
  var_data <- data.frame(Category = names(var_counts), Count = as.vector(var_counts))
  
  # Crear el gráfico de pastel
  pie_chart <- ggplot(var_data, aes(x = "", y = Count, fill = Category)) +
    geom_bar(stat = "identity", width = 1) +
    coord_polar(theta = "y") +
    theme_void() +
    geom_text(aes(label = paste0(round(Count/sum(Count) * 100, 1), "%")), position = position_stack(vjust = 0.5)) +
    scale_fill_manual(values = rainbow(length(var_counts))) +  # Cambia el esquema de color si deseas
    ggtitle(paste("Distribution of", var)) +
    theme(plot.title = element_text(hjust = 0.5))
  
  # Agregar el gráfico de pastel a la lista
  pie_charts[[var]] <- pie_chart
}

# Imprimir los gráficos de pastel
for (pie_chart in pie_charts) {
  print(pie_chart)
}

```

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Cmd+Shift+Enter*. 

```{r}

# Obtener nombres de las variables que son factores
factor_vars <- names(tfg)[sapply(tfg, is.factor)]

# Crear una lista para almacenar los gráficos de pastel
women_pie_charts <- list()
men_pie_charts <- list()

# Iterar sobre cada variable que es factor
for (var in factor_vars) {
  # Filtrar datos por género
  women_data <- tfg %>% filter(Gender == "Woman")
  men_data <- tfg %>% filter(Gender == "Man")
  
  # Contar las frecuencias de las categorías para mujeres y hombres
  women_var_counts <- table(women_data[[var]])
  men_var_counts <- table(men_data[[var]])
  
  # Crear dataframes con los datos
  women_var_data <- data.frame(Category = names(women_var_counts), Count = as.vector(women_var_counts))
  men_var_data <- data.frame(Category = names(men_var_counts), Count = as.vector(men_var_counts))
  
  # Crear los gráficos de pastel para mujeres y hombres
  women_pie_chart <- ggplot(women_var_data, aes(x = "", y = Count, fill = Category)) +
    geom_bar(stat = "identity", width = 1) +
    coord_polar(theta = "y") +
    theme_void() +
    geom_text(aes(label = paste0(round(Count/sum(Count) * 100, 1), "%")), position = position_stack(vjust = 0.5)) +
    scale_fill_manual(values = rainbow(length(women_var_counts))) +  # Cambia el esquema de color si deseas
    ggtitle(paste("Distribución de", var, "en mujeres")) +
    theme(plot.title = element_text(hjust = 0.5))
  
  men_pie_chart <- ggplot(men_var_data, aes(x = "", y = Count, fill = Category)) +
    geom_bar(stat = "identity", width = 1) +
    coord_polar(theta = "y") +
    theme_void() +
    geom_text(aes(label = paste0(round(Count/sum(Count) * 100, 1), "%")), position = position_stack(vjust = 0.5)) +
    scale_fill_manual(values = rainbow(length(men_var_counts))) +  # Cambia el esquema de color si deseas
    ggtitle(paste("Distribución de", var, "en hombres")) +
    theme(plot.title = element_text(hjust = 0.5))
  
  # Agregar los gráficos de pastel a las listas
  women_pie_charts[[var]] <- women_pie_chart
  men_pie_charts[[var]] <- men_pie_chart
}

# Imprimir los gráficos de pastel de mujeres
for (pie_chart in men_pie_charts) {
  print(pie_chart)
}

```
```{r}
# Imprimir los gráficos de pastel mujeres
for (pie_chart in women_pie_charts) {
  print(pie_chart)
}
```
```{r}
# Probabilidad de ser contratado
employed_prob <- sum(tfg$Employed == 1) / nrow(tfg)
women_employed_prob <- sum((tfg$Employed == 1)[tfg$Gender == 'Woman']) / sum(tfg$Gender == 'Woman')
men_employed_prob <- sum((tfg$Employed == 1)[tfg$Gender == 'Man']) / sum(tfg$Gender == 'Man')

# Display the probabilities
data.frame(
  Probability = c("P(Employed)", "P(Woman and Employed)", "P(Man and Employed)"),
  Probability_Value = c(employed_prob, women_employed_prob, men_employed_prob)
)
```
```{r}
# Create a loop to generate the bar plots
for (var in factor_vars) {
  bar_plots <- cc_barplot(Data = tfg, "Employed", var, freq = "relfreq")
  print(bar_plots)
}

```
```{r}
# Create a loop to generate the bar plots
for (var in factor_vars) {
  bar_plots <- cc_barplot(Data = tfg, var, "Employed", freq = "relfreq")
  print(bar_plots)
}
```
```{r}
hist(PreviousSalary, breaks = 20, col="lightblue", main="PreviousSalary", xlab="")
hist(ComputerSkills, breaks = 20, col="lightpink", main="ComputerSkills", xlab="")
hist(YearsCode, breaks = 20, col="orange", main="YearsCode", xlab="")
hist(YearsCodePro, breaks = 20, col="darkorange", main="YearsCodePro", xlab="")
```
```{r}
cc_hist(tfg,"PreviousSalary", "Employed", breaks=20)
cc_hist(tfg,"ComputerSkills", "Employed", breaks=20)
cc_hist(tfg,"YearsCode", "Employed", breaks=20)
cc_hist(tfg,"YearsCodePro", "Employed", breaks=20)
```
```{r}
cc_hist(tfg,"PreviousSalary", "MainBranch", breaks=20)
cc_hist(tfg,"PreviousSalary", "java", breaks=20)
cc_hist(tfg,"PreviousSalary", "SQL", breaks=20)
cc_hist(tfg,"PreviousSalary", "python", breaks=20)
cc_hist(tfg,"PreviousSalary", "c_plus_plus", breaks=20)
cc_hist(tfg,"PreviousSalary", "cloud", breaks=20)
```

```{r}
cc_hist(tfg,"PreviousSalary", "Gender", breaks=20)
cc_hist(tfg,"ComputerSkills", "Gender", breaks=20)
cc_hist(tfg,"YearsCode", "Gender", breaks=20)
cc_hist(tfg,"YearsCodePro", "Gender", breaks=20)
```
```{r}
plot(ComputerSkills, YearsCode, col=Employed)
plot(ComputerSkills, YearsCodePro, col=Employed)
```
```{r}
densityplot(~ PreviousSalary, data = tfg, groups = Continent, auto.key=TRUE)
densityplot(~ ComputerSkills, data = tfg, groups = Continent, auto.key=TRUE)
densityplot(~ ComputerSkills, data = tfg, groups = Gender, auto.key=TRUE)
densityplot(~ PreviousSalary, data = tfg, groups = EdLevel, auto.key=TRUE)
densityplot(~ ComputerSkills, data = tfg, groups = EdLevel, auto.key=TRUE)
```
```{r}
#identifing outliers 
boxplot(ComputerSkills, col = "green",notch = TRUE)
boxplot(PreviousSalary, col = "darkblue",notch = TRUE)
boxplot(YearsCode, col = "red",notch = TRUE)
```


```{r}
# Obtener nombres de las variables que son numericas
integer_vars <- names(tfg)[sapply(tfg, is.integer)]
integer_vars <- integer_vars[2:4]
str(integer_vars)


#CORRELACIONES ENTRE VARIABLES 
# convertir variables a numeros 
tfg1 <- tfg
for (var in factor_vars) {
  tfg1[, var] <- as.numeric(factor(tfg[, var]))-1
}

# Convert integer columns to numeric
for (var in integer_vars) {
  tfg1[, var] <- tfg[, var]
}

str(tfg1)
# Calculate the correlation matrix
tfg1$HaveWorkedWith <- NULL
tfg1$X <- NULL
tfg1$Country <- NULL
str(tfg1)

```


```{r}
#realizar t-test para valores numericos y chi2 para valores categoricos  
library(stats)
# Loop through the categorical variables and perform Chi-squared tests
T <- table(tfg$Gender,tfg$Employed)
T
#H0 --> no existe diferencia entre mujer y hombre 
chisq.test(T)
# p-value < 5%, null hypothesis can be rejected. significant association between the two variables.
```
```{r}
# Diferencia entre developers 
T <- table(tfg$MainBranch,tfg$Employed)
T
#H0 --> no existe diferencia
chisq.test(T)
# p-value < 5%, null hypothesis can be rejected. significant association between the two variables.
```

```{r}
# Diferencia entre developers 
T <- table(tfg$MainBranch,tfg$Gender)
T
#H0 --> no existe diferencia entre mujer y hombre 
chisq.test(T)
# p-value < 5%, null hypothesis can be rejected. significant association between the two variables.
```
```{r}
# Diferencia entre tech skills cloud  
T <- table(tfg$cloud,tfg$Gender)
T
#H0 --> no existe diferencia entre mujer y hombre 
chisq.test(T)
```

```{r}
# Diferencia entre tech skills cloud  
T <- table(tfg$cloud,tfg$Employed)
T
#H0 --> no existe diferencia entre mujer y hombre 
chisq.test(T)
```

```{r}
# Estimate linear regression
linear.model <- lm(tfg1$Employed ~ tfg1$Gender)
# Visualise data
plot( tfg1$Gender, tfg1$Employed)
# Visualise fitted linear regresion
abline(linear.model, col="red")
```
```{r}
# Estimate linear regression
linear.model <- lm(tfg1$Employed ~ tfg1$ComputerSkills)
# Visualise data
plot(tfg1$ComputerSkills, tfg1$Employed)
# Visualise fitted linear regresion
abline(linear.model, col="red")
```
```{r}
# Estimate linear regression
linear.model <- lm(tfg1$Employed ~ tfg1$MainBranch)
# Visualise data
plot(tfg1$MainBranch, tfg1$Employed)
# Visualise fitted linear regresion
abline(linear.model, col="red")
```
```{r}
#WEIGHTS AND INFORMATION VALUE 
require(Information)
# Create a Boolean variable that is TRUE when defaulted="No"
# (this choice is important as will be discussed later)
y <- tfg1$Employed == 0 #ESTAMOS CONSIDERANDO LA VARIABLE COMO NOT EMPLOYED 

# Multiplying by 1 renders y a numerical variable
y <- 1*y

# Creation/ addition of new variable called "class" in data.frame
tfg1["class"] <- y

# remove Employed
tfg1$Employed <- NULL
IV <- create_infotables(data=tfg1, y="class", bins = 5)
IV
#Odds positivos significa que los odds de no ser contratado siendo mujer es mayor a los odds de la poblacion 
#mientras que los odds negativos significan que eres mas probable en ser contratado si eres hombre 
```
```{r}
#WEIGHTS AND INFORMATION VALUE 
require(Information)
# Create a Boolean variable that is TRUE when defaulted="No"
# (this choice is important as will be discussed later)
y <- tfg1$class == 1 #ESTAMOS CONSIDERANDO LA VARIABLE EMPLOYED (CONTRATADO)

# Multiplying by 1 renders y a numerical variable
y <- 1*y

# Creation/ addition of new variable called "class" in data.frame
tfg1["class"] <- y

# remove Employed
tfg1$Employed <- NULL
IV <- create_infotables(data=tfg1, y="class", bins = 5)
IV
#Odds positivos significa que los odds de no ser contratado siendo mujer es mayor a los odds de la poblacion 
#mientras que los odds negativos significan que eres mas probable en ser contratado si eres hombre 
```

```{r}
# funcion para eliminar outliers en base al IQR 
remove_outliers_iqr <- function(data, column_name) {
  Q1 <- quantile(data[[column_name]], 0.25)
  Q3 <- quantile(data[[column_name]], 0.75)
  IQR <- Q3 - Q1
  lower_bound <- Q1 - 1.5 * IQR
  upper_bound <- Q3 + 1.5 * IQR
  data <- data[data[[column_name]] >= lower_bound & data[[column_name]] <= upper_bound, ]
  return(data)
}

# Remove outliers in 'YearsCode', 'PreviousSalary,' and 'ComputerSkills'
tfg1 <- remove_outliers_iqr(tfg1, "YearsCode")
tfg1 <- remove_outliers_iqr(tfg1, "PreviousSalary")
tfg1 <- remove_outliers_iqr(tfg1, "ComputerSkills")
#OBSERVACIONES = 66958

```

```{r}
#setwd("tfg/Datos y formulas")
write.csv(tfg1, "datos_logit.csv", row.names = FALSE)
```
